+++ yt/core/misc/error.cpp
@@ -248,8 +248,8 @@ void AppendError(TStringBuilder* builder, const TError& error, int indent)
             case ETokenType::String:
                 AppendAttribute(builder, key, Stroka(tokenizer.CurrentToken().GetStringValue()), indent);
                 break;
-            case ETokenType::Integer:
-                AppendAttribute(builder, key, ToString(tokenizer.CurrentToken().GetIntegerValue()), indent);
+            case ETokenType::Int64:
+                AppendAttribute(builder, key, ToString(tokenizer.CurrentToken().GetInt64Value()), indent);
                 break;
             case ETokenType::Double:
                 AppendAttribute(builder, key, ToString(tokenizer.CurrentToken().GetDoubleValue()), indent);
